+++ Source/Core/WindingOrder.js
@@ -1,5 +1,8 @@
 /*global define*/
-define(function() {
+define([
+        './freezeObject'
+    ], function(
+        freezeObject) {
     "use strict";
 
     /**
@@ -32,5 +35,5 @@ define(function() {
         }
     };
 
-    return WindingOrder;
+    return freezeObject(WindingOrder);
 });