+++ Source/Scene/StencilFunction.js
@@ -1,5 +1,8 @@
 /*global define*/
-define(function() {
+define([
+        '../Core/freezeObject'
+    ], function(
+        freezeObject) {
     "use strict";
 
     /**
@@ -73,5 +76,5 @@ define(function() {
         ALWAYS : 0x0207
     };
 
-    return StencilFunction;
+    return freezeObject(StencilFunction);
 });